rename /routes/members.js to /routes/teamMembers.js
