+++ src/routes/index.js
@@ -14,7 +14,7 @@ import accountMembersRoutes from './accountMembers';
 import authRoutes from './auth';
 import teamsRoutes from './teams';
 import teamAccountsRoutes from './teamAccounts';
-import teamMembersRoutes from './members';
+import teamMembersRoutes from './teamMembers';
 import usersRoutes from './users';
 
 /**
